policy

+++ app/Http/Controllers/TestimonialController.php
@@ -19,6 +19,8 @@ class TestimonialController extends Controller
     }
     public function store(Request $request)
     {
+        $this->authorize("create", Testimonial::class);
+
         $testimonial = new Testimonial;
         $request->validate([
             'name' => 'required',
@@ -44,6 +46,7 @@ class TestimonialController extends Controller
     public function update($id, Request $request)
     {
         $testimonial = Testimonial::find($id);
+        $this->authorize("update", $testimonial);
         $request->validate([
             'name' => 'required',
             'post' => 'required',
@@ -58,6 +61,7 @@ class TestimonialController extends Controller
     public function destroy($id)
     {
         $testimonial = Testimonial::find($id);
+        $this->authorize("delete", $testimonial);
         $testimonial->delete();
         return redirect()->back()->with("message", "Successful delete !");
     }
